requirements.txt
flask
requests
moviepy
imageio-ffmpeg
rembg
pillow
onnxruntime